requirements.txt
Django==5.2.8
pandas==2.3.3
python_dateutil==2.9.0.post0
